--- src/main/groovy/io/hhplus/tdd/hhplusconcertjava/point/apllication/PointEventListener.java
@@ -1,6 +1,0 @@
-package io.hhplus.tdd.hhplusconcertjava.point.apllication;
-
-public interface PointEventListener {
-
-    public void useCancel(UseCancelEvent useCancelEvent);
-}
