+++ src/main/groovy/io/hhplus/tdd/hhplusconcertjava/point/domain/entity/UseCancelEvent.java
@@ -1,0 +1,10 @@
+package io.hhplus.tdd.hhplusconcertjava.point.domain.entity;
+
+import lombok.Builder;
+import lombok.Data;
+
+@Data
+@Builder
+public class UseCancelEvent {
+    private PointHistory pointHistory;
+}
