+++ src/main/groovy/io/hhplus/tdd/hhplusconcertjava/point/infrastructure/PointEventListener.java
@@ -1,0 +1,8 @@
+package io.hhplus.tdd.hhplusconcertjava.point.infrastructure;
+
+import io.hhplus.tdd.hhplusconcertjava.point.domain.entity.UseCancelEvent;
+
+public interface PointEventListener {
+
+    public void useCancel(UseCancelEvent useCancelEvent);
+}
