--- src/main/groovy/io/hhplus/tdd/hhplusconcertjava/wait/application/WaitEventListener.java
@@ -1,5 +1,0 @@
-package io.hhplus.tdd.hhplusconcertjava.wait.application;
-
-public interface WaitEventListener {
-    public void deleteActivateToken(DeleteActivateTokenEvent deleteActivateTokenEvent);
-}
