+++ src/main/groovy/io/hhplus/tdd/hhplusconcertjava/wait/domain/entity/DeleteActivateTokenEvent.java
@@ -1,0 +1,10 @@
+package io.hhplus.tdd.hhplusconcertjava.wait.domain.entity;
+
+import lombok.Builder;
+import lombok.Data;
+
+@Data
+@Builder
+public class DeleteActivateTokenEvent {
+    String uuid;
+}
