+++ src/main/groovy/io/hhplus/tdd/hhplusconcertjava/wait/infrastructure/WaitEventListener.java
@@ -1,0 +1,7 @@
+package io.hhplus.tdd.hhplusconcertjava.wait.infrastructure;
+
+import io.hhplus.tdd.hhplusconcertjava.wait.domain.entity.DeleteActivateTokenEvent;
+
+public interface WaitEventListener {
+    public void deleteActivateToken(DeleteActivateTokenEvent deleteActivateTokenEvent);
+}
