+++ src/main/groovy/io/hhplus/tdd/hhplusconcertjava/wait/infrastructure/WaitTokenRedisRepository.java
@@ -12,7 +12,6 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 import java.util.Set;
 import java.util.UUID;
-import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.atomic.AtomicLong;
 
 @Slf4j
