+++ src/test/groovy/io/hhplus/tdd/hhplusconcertjava/integration/wait/TestWaitEvent.java
@@ -1,7 +1,7 @@
 package io.hhplus.tdd.hhplusconcertjava.integration.wait;
 
 import io.hhplus.tdd.hhplusconcertjava.integration.TestBaseIntegration;
-import io.hhplus.tdd.hhplusconcertjava.wait.application.DeleteActivateTokenEvent;
+import io.hhplus.tdd.hhplusconcertjava.wait.domain.entity.DeleteActivateTokenEvent;
 import io.hhplus.tdd.hhplusconcertjava.wait.domain.entity.ActivateToken;
 import io.hhplus.tdd.hhplusconcertjava.wait.domain.repository.ActivateTokenRepository;
 import org.junit.jupiter.api.Nested;
@@ -10,9 +10,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.ApplicationEventPublisher;
-import org.springframework.transaction.annotation.Transactional;
 
-import java.util.Random;
 import java.util.UUID;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
